added support for the "public network" (bridged)
When you are using the default NAT network, "apt-get dist-upgrade" will fail if your host's antivirus or endpoint protection blocks one of the packages.  The workaround is to choose a bridged network to bypass endpoint protection/antivirus.   I added support so that Vagrant doesn't break if you choose to configure a "public network" through the vagrantfile (rather than manually configuring a bridged network directly in virtualbox).  Read comments in this file for more detail.